+++ lib/l10n/translations/lt.dart
@@ -1791,6 +1791,9 @@ const _dutchSourceLt = {
 };
 
 const _dutchSourceAddLt = <String, String>{
+  'Deze presentatie bevat onderdelen van de Informatieveiligheidsmodule. Zet de module aan om ze te bewerken.':
+      'Šioje pateiktyje yra Informacijos saugumo modulio dalių. Įjunkite modulį, kad galėtumėte jas redaguoti.',
+  'Inschakelen': 'Įjungti',
   'Bevestigen': 'Patvirtinti',
   'Onderbouwing van de bevestiging': 'Patvirtinimo pagrindimas',
   'De eigenaar staat insluiten niet toe': 'Savininkas neleidžia įterpti',