+++ lib/l10n/translations/lv.dart
@@ -1782,6 +1782,9 @@ const _dutchSourceLv = {
 };
 
 const _dutchSourceAddLv = <String, String>{
+  'Deze presentatie bevat onderdelen van de Informatieveiligheidsmodule. Zet de module aan om ze te bewerken.':
+      'Šī prezentācija satur Informācijas drošības moduļa daļas. Ieslēdziet moduli, lai tās rediģētu.',
+  'Inschakelen': 'Ieslēgt',
   'Bevestigen': 'Apstiprināt',
   'Onderbouwing van de bevestiging': 'Apstiprinājuma pamatojums',
   'De eigenaar staat insluiten niet toe': 'Īpašnieks neatļauj iegulšanu',