+++ lib/l10n/translations/tlh.dart
@@ -1758,6 +1758,9 @@ const _dutchSourceTlh = {
 };
 
 const _dutchSourceAddTlh = <String, String>{
+  'Deze presentatie bevat onderdelen van de Informatieveiligheidsmodule. Zet de module aan om ze te bewerken.':
+      'nav vam De\' Qan module Dochmey ngaS. Dochmeyvam yIchoHmeH module yIchu\'.',
+  'Inschakelen': 'yIchu\'',
   'Bevestigen': 'wuq',
   'Onderbouwing van de bevestiging': 'meq',
   'De eigenaar staat insluiten niet toe': 'ghajwI\' chaw\'be\'',